+++ Greedy/src/main/scala/com/gjos/scala/swoc/protocol/Stone.scala
@@ -1,23 +1,23 @@
 package com.gjos.scala.swoc.protocol
 
 trait Stone {
-  val value: Byte
+  val value: Int
 }
 
 object Stone {
   case object Pebble extends Stone {
-    val value: Byte = 1
+    val value: Int = 1
   }
 
   case object Rock extends Stone {
-    val value: Byte = 2
+    val value: Int = 2
   }
 
   case object Boulder extends Stone {
-    val value: Byte = 3
+    val value: Int = 3
   }
 
-  def byValue(i: Byte): Option[Stone] = i match {
+  def byValue(i: Int): Option[Stone] = i match {
     case 0 => None
     case 1 => Some(Pebble)
     case 2 => Some(Rock)
